Show password reset as modal on home
